# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: GrandGold Marketplace API
  description: |
    GrandGold is a multi-country luxury jewellery marketplace with support for India (INR), UAE (AED), and UK (GBP).
    
    ## Authentication
    Most endpoints require Bearer token authentication. Include the JWT token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    - 200 requests per minute for product APIs
    - 100 requests per minute for order/payment APIs
    
    ## Countries
    Supported countries: `IN` (India), `AE` (UAE), `UK` (United Kingdom)
  version: 1.0.0
  contact:
    name: GrandGold API Support
    email: api@thegrandgold.com
  license:
    name: Proprietary
    url: https://thegrandgold.com/terms
servers:
  - url: https://api.thegrandgold.com
    description: Production
  - url: https://staging-api.thegrandgold.com
    description: Staging
  - url: http://localhost:3000
    description: Local Development
...
paths:
...
  /api/products:
    get:
      tags: [Products]
      summary: List products
      operationId: listProducts
      parameters:
        - name: country
          in: query
          schema:
            type: string
            enum: [IN, AE, UK]
        - name: category
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Product list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  total:
                    type: integer
  /api/products/{id}:
    get:
      tags: [Products]
      summary: Get product by ID
      operationId: getProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        sku:
          type: string
        price:
          type: number
        currency:
          type: string
        category:
          type: string
        subcategory:
          type: string
        purity:
          type: string
          enum: ['14K', '18K', '22K', '24K']
        weight:
          type: number
        images:
          type: array
          items:
            type: string
        sellerId:
          type: string
        stock:
          type: integer
        isActive:
          type: boolean
        country:
          type: string
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
            code:
              type: string
            details:
              type: object
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...